refactor: Export computeHomeApp from HomeLinker to helpers
to facilitate mocking in tests
